moved auth0.js out of pages folder due to nextjs version 9 updates
