+++ periodo/package.json
@@ -8,16 +8,18 @@
     "start": "next start"
   },
   "dependencies": {
+    "@ampproject/toolbox-optimizer": "^2.7.6",
     "@auth0/nextjs-auth0": "^1.2.0",
     "@emotion/core": "^11.0.0",
     "@emotion/react": "^11.1.4",
     "@emotion/styled": "^11.0.0",
     "@fullcalendar/react": "^5.5.0",
     "@material-ui/core": "^4.11.2",
-    "babel-plugin-transform-define": "^2.0.0",
-    "css-loader": "^5.1.3",
     "@material-ui/icons": "^4.11.2",
+    "babel-plugin-transform-define": "^2.0.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "beauty-stars": "^1.1.0",
+    "css-loader": "^5.1.3",
     "dotenv": "^8.2.0",
     "firebase": "^8.2.3",
     "next": "^10.0.8",
@@ -33,6 +35,7 @@
     "react-star-rating-component": "^1.4.1",
     "resolve-url-loader": "^3.1.2",
     "sass-loader": "^11.0.1",
+    "sharp": "^0.27.2",
     "style-loader": "^2.0.0",
     "webpack": "^5.27.2"
   },
