+++ periodo/pages/api/auth/login.js
@@ -1,5 +1,5 @@
 
-import auth0 from '../../utils/auth0';
+import auth0 from '../../../utils/auth0';
 
 export default async function login(req, res) {
   try {
