+++ periodo/pages/api/auth/logout.js
@@ -1,4 +1,4 @@
-import auth0 from '../../utils/auth0';
+import auth0 from '../../../utils/auth0';
 export default async function logout(req, res) {
     try {
         await auth0.handleLogout(req, res);
