+++ periodo/tsconfig.json
@@ -22,7 +22,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-, "pages/_app.js", "pages/_document.js", "pages/utils/auth0.js", "pages/api/auth/callback.js"  ],
+, "pages/_app.js", "pages/_document.js", "utils/auth0.js", "pages/api/auth/callback.js"  ],
   "exclude": [
     "node_modules"
   ]
